Improve error message for catalog errors when listing metadata
Depending on the client that initiated a metadata query, error messages
returned from `ConnectorMetadata` can be confusing for users. For
example, a query to `system.jdbc` tables with no catalog predicate, on
exceptions in the catalog, will propagate those exceptions, with no way
for the user to know which catalog has thrown that exception.

When there are a lot of catalogs, this makes investigating the issue
very hard for Trino admins. We can make it easier by always including
the problematic catalog's name in the error message.

(cherry picked from commit e9b931f500d1dee968cee81ded4ae0f86ee4158e)