Log impacted features
This is helpful in CI to understand the context on how tests are
executed.

Impacted features are passed as a log file and then they are passed from
suite to test run, so logging them here make sure if everything is
passed correctly.

(cherry picked from commit 58d635621b63123cedbeb02557ddac617ab32337)