Introduce EmptyTableFunctionHandle as default handle
Before this change, if a table function did not pass
a ConnectorTableFunctionHandle in the TableFunctionAnalysis,
the default handle was used, which was an anonymous
implementation of ConnectorTableFunctionHandle.

It did not work with table functions executed by operator,
due to lack of serialization.

This change introduces EmptyTableFunctionHandle, and sets
it as default.

(cherry picked from commit 0efebddba517d4ba2fafa5792283b7b3d5ccd2ce)