Tune JDBC fetch-size automatically based on column count
PostgreSQL, Redshift and Oracle connectors had hard-coded fetch-size
value of 1000. The value was found not to be optimal when server is far
(high latency) or when number of columns selected is low. This commit
improves in the latter case by picking fetch size automatically based on
number of columns projected. After the change, the fetch size will be
automatically picked in the range 1000 to 100,000.

(cherry picked from commit 9c4ef0e98606302dd593072145f9a75d967cfecf)