Pass partition values as optional to DeltaLakePageSource
The partition values list is filled only when row ID column is
projected, so it's a conditional information. When row ID is not
present, pass it as the empty optional, rather than list that happens to
be empty.

(cherry picked from commit 049b1d42b810ca616a0bb4bc203216a59c608a94)