Hoist a creating MethodHandle to PageProjection constructor
For every created PageProjection engine creates a MethodHandle
to a PageProjectionWork constructor, even when it is cached.
Creating (possibly many) MethodHandles for every query is not
a correct approach so if it can be avoided, it should be avoided.

(cherry picked from commit c9309abcf9b71bd56dc337846993c99d40540299)