Refactor internal http client usage
The previous approach was dangerous, as it was modifying configuration for all HTTP clients and adding global filters even for HTTP clients there were not
supposed to be used for internal communication.

With this change, it's now intentional which HTTP clients are supposed to use secured internal communication.

(cherry picked from commit 6c8dd6ebfc24741746b0692045198aba5d6af6f7)