Provide proper Trino version for SPI compatibility checks
We use revapi maven plugin to check SPI backward compatibility.
By default, revapi compares current code to the latest released
version of the artifact. This causes CI failures if starburst-trino
is at VERSION-e-SNAPSHOT and Trino VERSION has been already
released. In such case, we perform "reverted check" (VERSION-1
vs VERSION) and it fails for any API changes introduced in VERSION.

We need to explicitly set the proper version to compare to. It
should always be the previous Trino version (VERSION-1).