Reduce memory usage in TestRubixCaching
Previously the test wouldn't pass with heap lower than 900 MB and would
be flaky with a heap somewhat bigger than that. This is because it
materializes a 100 MB "large file" in memory a few times.

This change reduces number of instantiations and makes the test pass
with 200 MB heap.

(cherry picked from commit cc4884a1693e65e25f6120caf72ed5a0f4831c65)