Fix DEFAULT_ON_NULL return convention to short circuit on null argument
The DEFAULT_ON_NULL return convention should not cause the function to
be invoked when there is a null passed to a never null argument.
Instead, as with NULLABLE_RETURN, the invocation should be
short- circuited and the default value should be returned.

(cherry picked from commit 64ada5c950b522196cc5321cf75268cd3b48142b)