Simplify ChannelSet and GroupByHash contains
ChannelSet is hard coded to pass a simple page with the values in
channel 0 and an optional precomputed hash in channel 1. This aligns
with the assumptions of the GroupByHash, so the hashChannels argument to
the contains method is not needed.

(cherry picked from commit 02dc79686562d65203d200dafc1de1e4178d20d9)