Run all the tests, except FTE, by default on `trino-redshift` module
The project `starburst-trino` benefits of dedicated CI/CD
infrastructure, with the runners placed on a specific availability zone.
This aspect makes `starburst-trino` a good candidate to run all the tests
on the `trino-redshift` module, because the runtime of the tests will not
vary based on the random geographical location of the Azure Github runner.

By ensuring that all the tests on `trino-redshift` run accurately, the
module's code can be safely reused downstream in `starburst-enterprise`.