Refactor KafkaCatalog in Kafka product tests
The message serializer being used is strictly coupled to the catalog
being used so the serializer also belongs in the KafkaCatalog. This is
useful for an upcoming commit to be able to simplify the usages of
DataProvider in some product tests.

(cherry picked from commit 9f9393a08c6aec1f0b422b4bf301f565d192c4e3)