Add product tests for Kafka connector with Confluent licensed libraries
The older approach was to copy the Confluent licensed libraries to the
Kafka connector classpath for all product environments extending from
`Kafka`. This meant that it was not possible to detect changes which
introduce a hard-dependency on the Confluent licensed libraries for
Kafka connector startup.

02cc332319a00e8fec5096921f96cb9e0aceacc1 inadvertently made it so that
the Confluent libraries became required for the Kafka connector to start
up and was not caught by tests because all the product test environments
already had the Confluent libraries present on the Kafka connector
classpath.

This commit instead introduces a new product test environment
`EnvMultinodeKafkaConfluentLicense` which is the only product test
environment that copies the Confluent licensed libraries onto the Kafka
connector classpath. There is also a new test group
`kafka_confluent_license` which now includes all the tests for the
functionality which requires the Confluent licensed libraries.

(cherry picked from commit a1c8d8ef960606637ba53f2b00fea74a77553a79)