Improve feedback from CI runs with all secrets
When starting the CI run with all secrets, add a comment to the original
PR with a link to that run, instead of only adding a comment after such
run would fail. Update the comment with the status once it finishes.

(cherry picked from commit 1882c1dafe8abd75485899f6eebfddd7fe893abb)