Fix env not to use secret unnecessarily
`EnvSinglenodeDeltaLakeOss` was not working when PR's CI build is
triggered from the fork. The secrets and variables are not present, and
this resulted in `S3_BUCKET` env variable being defined but empty.
However, the environment doesn't need to be parameterized with the
bucket name at all.

(cherry picked from commit 4c057fb46bbf1b017265ad436b10340a3d1e5290)