Remove broken memory connector test
The test attemps to ensure that queries will fail
if one of the nodes holding a partition of the data
goes offline. However, it's fundamentally broken,
as it doesn't do anything to ensure that the node
that's being shut down actually contains any data
for the table in question.

With the recent changes in https://github.com/trinodb/trino/pull/18005
that enable writer scaling for any connector, it
is very likely that the table will land in a subset
of the nodes and cause the test to fail if the "wrong"
node is shut down.

(cherry picked from commit c2427c725ba314019b15587b42b5de15fcec2d62)