Fix race condition in getTopRuleStats
The code is sorting with a comparator that's based on the mutable
state of QueryPlanOptimizerStats. When the state changes in the
middle of the sort, it can cause the sort to behave unexpectedly
(hang, fail, produce incorrect results, etc).

(cherry picked from commit 8de71d556418863e3cc7433a5e6d4568ff7689fd)