Fix check in HiveMetadata.getView and update its name
The method is supposed to get a Trino view that's going to be updated.
Its check for `translateHiveViews` was redundant, the calling code is
never supposed to operate on a Hive view, regardless of
`translateHiveViews` flag. (And if it matter, the `translateHiveViews`
condition would probably be negated.) It was probably not user-facing
problem as the calling code is likely unreachable when Hive view
translation is not enabled.

The commit also adds a check that the table object being operated on
does not represent a Trino Materialized View. The previously checked
`isPrestoView` returns true for both, but the table is supposed to
represent ordinary Trino View.

(cherry picked from commit 65cbdcd93fa10f34739c35d72c5fb9a24f1f9c7b)